Implement code review suggestions.
