+++ Core/Core/Common/CommonModels/Analytics/DebugDescriptionProvider.swift
@@ -20,15 +20,3 @@ public protocol DebugDescriptionProvider {
     /// The non-localized string representation of an entity for debugging / analytics purposes.
     var debugDescription: String { get }
 }
-
-extension Error {
-
-    /// Convenience method to get the `debugDescription` from the error.
-    /// If the error is not a `DebugDescriptionProvider`, it will return its localized description.
-    public var debugDescription: String {
-        guard let debugDescription = (self as? DebugDescriptionProvider)?.debugDescription else {
-            return localizedDescription
-        }
-        return debugDescription
-    }
-}
